MAINT: refactor, centralize setup boilerplate
